Fix some win64 runtime issues with time_t and file handles

+++ fs2_p4_main/code/CFile/cfile.cpp
@@ -318,7 +318,7 @@ int cfile_pop_dir()
 // NOTE : WILL NOT DELETE READ-ONLY FILES
 int cfile_flush_dir(int dir_type)
 {
-	int find_handle;
+	intptr_t find_handle;
 	int del_count;
 	_finddata_t find;
