+++ fs2_p4_main/code/FREESPACE2/FreeSpace.cpp
@@ -7012,7 +7012,7 @@ int find_freespace_cd(char *volume_name)
 	int cdrom_drive=-1;
 	int volume_match = 0;
 	_finddata_t find;
-	int find_handle;
+	intptr_t find_handle;
 
 	GetCurrentDirectory(MAX_PATH, oldpath);
